+++ catalyst/core/callbacks/metrics.py
@@ -1,15 +1,13 @@
-from typing import (  # isort:skip
-    Any, Callable, List, Dict, Union, TYPE_CHECKING  # isort:skip
-)  # isort:skip
+from collections import defaultdict
 from functools import partial
 import logging
+from typing import Any, Callable, Dict, List, Union
 
 import torch
 
-from catalyst.core import Callback, CallbackOrder
-
-if TYPE_CHECKING:
-    from catalyst.core import _State
+from catalyst import utils
+from catalyst.core import _State, Callback, CallbackNode, CallbackOrder
+from catalyst.utils import meters
 
 logger = logging.getLogger(__name__)
 
@@ -32,7 +30,7 @@ class MetricCallback(Callback):
         self.input_key = input_key
         self.output_key = output_key
 
-    def on_batch_end(self, state: "_State"):
+    def on_batch_end(self, state: _State):
         outputs = state.batch_out[self.output_key]
         targets = state.batch_in[self.input_key]
         metric = self.metric_fn(outputs, targets)
@@ -59,7 +57,7 @@ class MultiMetricCallback(Callback):
         self.input_key = input_key
         self.output_key = output_key
 
-    def on_batch_end(self, state: "_State"):
+    def on_batch_end(self, state: _State):
         outputs = state.batch_out[self.output_key]
         targets = state.batch_in[self.input_key]
 
@@ -163,10 +161,66 @@ class MetricAggregatorCallback(Callback):
 
         return result
 
-    def on_batch_end(self, state: "_State") -> None:
+    def on_batch_end(self, state: _State) -> None:
         """
         Computes the metric and add it to the metrics
         """
         metrics = self._preprocess(state.batch_metrics)
         metric = self.aggregation_fn(metrics)
         state.batch_metrics[self.prefix] = metric
+
+
+class MetricsManagerCallback(Callback):
+    """
+    Prepares metrics for logging, transferring values from PyTorch to numpy
+    """
+    def __init__(self):
+        super().__init__(
+            order=CallbackOrder.Logging - 1,
+            node=CallbackNode.All,
+        )
+        self.meters: Dict[str, meters.AverageValueMeter] = None
+
+    @staticmethod
+    def _to_single_value(value: Any) -> float:
+        if hasattr(value, "item"):
+            value = value.item()
+
+        value = float(value)
+        return value
+
+    @staticmethod
+    def _process_metrics(metrics: Dict[str, Any]):
+        output = {}
+        for key, value in metrics.items():
+            value = MetricsManagerCallback._to_single_value(value)
+            value = utils.distributed_mean(value)
+            output[key] = value
+        return output
+
+    def on_epoch_start(self, state: _State):
+        state.epoch_metrics = defaultdict(None)
+
+    def on_loader_start(self, state: _State):
+        state.loader_metrics = defaultdict(None)
+        self.meters = defaultdict(meters.AverageValueMeter)
+
+    def on_loader_end(self, state: _State):
+        for key, value in self.meters.items():
+            value = value.mean
+            state.loader_metrics[key] = value
+            state.epoch_metrics[f"{state.loader_name}_{key}"] = value
+
+    def on_batch_start(self, state: _State):
+        state.batch_metrics = defaultdict(None)
+
+    def on_batch_end(self, state: _State):
+        state.batch_metrics = self._process_metrics(state.batch_metrics)
+        for key, value in state.batch_metrics.items():
+            self.meters[key].add(value)
+
+
+__all__ = [
+    "MetricCallback", "MultiMetricCallback", "MetricAggregatorCallback",
+    "MetricsManagerCallback"
+]
